WIP starting on unified AST
Learned so far: we might want to wait for GHC 8.2, since I think it
will support COMPLETE pragmas for pattern matching with pattern
synonyms.

There's going to be a lot of boilerplate in the module that defines
the base AST, but hopefully the other modules defining variants of it
for various phases will be not too bad.